Uploads .dae files
Rather than .kmz.
This makes the use of double_sided flag possible and uploading a correct model.
Thus fixing the culling bug of faces by their backfaces.

To group .dae and textures, the files are zipping using a command line tool.
For windows it was necessary to embed the binary 7za.exe.
For other OSes the zip command is used.